Implemented twitter api call
Now backend calls twitter API to search for a twitter username on each of the politicians found. Then, the data from congress and twitter are combined and saved to a mongodb instance (also configured with schema in this commit). If the politicians exists in complete form in the database, it is returned to the screen. If not, politician object is completed with twitter username information and persisted to the db.

Also improved architecture on node.js backend.