Added twitter timeline
Tried my best to push the timeline's width to be the size of the window, but it turns out it was not made to be large... even on twitter.com it is pretty thin. Font size and media size get heavily outscaled when the window is fullscreen on max-width. I will reconsider and grid the politicians page instead.